Remove io.directories module
