+++ gpm_api/io/pps.py
@@ -11,12 +11,177 @@ import subprocess
 from dateutil.relativedelta import relativedelta
 
 from gpm_api.configs import get_gpm_password, get_gpm_username
-from gpm_api.io.directories import (
-    get_pps_directory,
-    get_pps_directory_tree,
-    get_pps_servers,
+from gpm_api.io.checks import (
+    check_product_type,
+    check_product_validity,
+    check_product_version,
 )
 from gpm_api.io.products import available_products, get_info_dict
+
+
+def _get_pps_nrt_product_folder_name(product):
+    """ "Retrieve NASA PPS server folder name for NRT product_type."""
+    folder_name = get_info_dict()[product].get("pps_nrt_dir", None)
+    if folder_name is None:
+        raise ValueError(
+            f"The pps_nrt_dir key of the {product} product is not specified in the config files."
+        )
+    return folder_name
+
+
+def _get_pps_rs_product_folder_name(product):
+    """ "Retrieve NASA PPS server folder name for RS product_type."""
+    folder_name = get_info_dict()[product].get("pps_rs_dir", None)
+    if folder_name is None:
+        raise ValueError(
+            f"The pps_rs_dir key of the {product} product is not specified in the config files."
+        )
+    return folder_name
+
+
+def _get_pps_nrt_product_dir(product, date):
+    """
+    Retrieve the NASA PPS server directory structure where NRT data are stored.
+
+    Parameters
+    ----------
+    product : str
+        GPM product name. See: gpm_api.available_products() .
+    date : datetime.date
+        Single date for which to retrieve the data.
+        Note: this is currently only needed when retrieving IMERG data.
+    """
+    folder_name = _get_pps_nrt_product_folder_name(product)
+    # Specify the directory tree
+    if product in available_products(product_type="NRT", product_category="IMERG"):
+        directory_tree = os.path.join(folder_name, datetime.datetime.strftime(date, "%Y%m"))
+    else:
+        directory_tree = folder_name
+    return directory_tree
+
+
+def _get_pps_rs_product_dir(product, date, version):
+    """
+    Retrieve the NASA PPS server directory structure where RS data are stored.
+
+    Parameters
+    ----------
+    product : str
+        GPM product name. See: gpm_api.available_products() .
+
+    date : datetime.date
+        Single date for which to retrieve the data.
+    version : int, optional
+        GPM version of the data to retrieve if product_type = 'RS'.
+    """
+    version = check_product_version(version, product)
+    check_product_validity(product, product_type="RS")
+
+    # Retrieve NASA server folder name for RS
+    folder_name = _get_pps_rs_product_folder_name(product)
+
+    # Specify the directory tree for current RS version
+    if version == 7:
+        directory_tree = os.path.join(
+            "gpmdata",
+            datetime.datetime.strftime(date, "%Y/%m/%d"),
+            folder_name,
+        )
+
+    # Specify the directory tree for old RS version
+    else:  #  version in [4, 5, 6]:
+        version_str = "V0" + str(int(version))
+        directory_tree = os.path.join(
+            "gpmallversions",
+            version_str,
+            datetime.datetime.strftime(date, "%Y/%m/%d"),
+            folder_name,
+        )
+
+    # Return the directory tree
+    return directory_tree
+
+
+def get_pps_directory_tree(product, product_type, date, version):
+    """
+    Retrieve the NASA PPS server directory tree where the GPM data are stored.
+
+    Parameters
+    ----------
+    product : str
+        GPM product name. See: gpm_api.available_products() .
+    product_type : str, optional
+        GPM product type. Either 'RS' (Research) or 'NRT' (Near-Real-Time).
+    date : datetime.date
+        Single date for which to retrieve the data.
+    version : int, optional
+        GPM version of the data to retrieve if product_type = 'RS'.
+
+    Returns
+    -------
+    directory_tree : str
+        DIrectory tree on the NASA PPS server where the data are stored.
+    """
+    check_product_type(product_type)
+    if product_type == "NRT":
+        return _get_pps_nrt_product_dir(product, date)
+    else:  # product_type == "RS"
+        return _get_pps_rs_product_dir(product, date, version)
+
+
+def get_pps_servers(product_type):
+    """Return the url to the PPS servers."""
+    if product_type == "NRT":
+        url_text_server = "https://jsimpsonhttps.pps.eosdis.nasa.gov/text"
+        url_data_server = "ftps://jsimpsonftps.pps.eosdis.nasa.gov/data"
+    else:
+        url_text_server = "https://arthurhouhttps.pps.eosdis.nasa.gov/text"
+        url_data_server = "ftps://arthurhouftps.pps.eosdis.nasa.gov"
+    return (url_text_server, url_data_server)
+
+
+def get_pps_directory(product, product_type, date, version):
+    """
+    Retrieve the NASA PPS server directory paths where the GPM data for
+    a specific date are listed and stored.
+
+    The data list is retrieved using https.
+    The data stored are retrieved using ftps.
+
+    Parameters
+    ----------
+    product : str
+        GPM product name. See: gpm_api.available_products() .
+    product_type : str, optional
+        GPM product type. Either 'RS' (Research) or 'NRT' (Near-Real-Time).
+    date : datetime.date
+        Single date for which to retrieve the data.
+    version : int, optional
+        GPM version of the data to retrieve if product_type = 'RS'.
+
+    Returns
+    -------
+    url_data_server : str
+        url of the NASA PPS server where the data are stored.
+    url_data_list: list
+        url of the NASA PPS server where the data are listed.
+
+    """
+    check_product_type(product_type)
+
+    # Retrieve servers URLs
+    url_text_server, url_data_server = get_pps_servers(product_type)
+
+    # Retrieve directory tree structure
+    dir_structure = get_pps_directory_tree(
+        product=product, product_type=product_type, date=date, version=version
+    )
+
+    # Define url where data are listed
+    url_data_list = os.path.join(url_text_server, dir_structure)
+
+    # Return tuple
+    return (url_data_server, url_data_list)
 
 
 def ensure_valid_start_date(start_date, product):
